cache_access, update_policy, and setup_cache work correctly
cache_access is changed so that it makes separate calls to the
read and write functions when multipl blocks are being read or
written. It also calculates the tag, index, and offset now.

update_policy only works for LRU for now. It is working how it is
supposed to.

setup_cache has been working but now it works for sure. It will
create caches correctly.